split chatView file
